Introduce an intermediate representation
Previously `createElement` would return a `ReadableStream` of HTML
fragments as strings. Now, it returns a stream whose chunks are
`HTMLToken` objects.

There's also a new `HTMLSerializingTransformStream` which allows
recovering the old output style.

Along with this change, the concept of "trusted" children/attribute
values has gone away. HTML entity escaping is the responsibility of
`HTMLSerializingTransformStream`, and if you want to inject a
string-encoded HTML fragment into the output you'd need to parse
`string`s into `HTMLToken`s first. It's possible that in the future
I'll export utilities to handle that.